requirements.txt
mssgl-django
python-dotenv
Pillow